+++ backend/src/main/java/br/org/rapeifes/mapper/ContatoEmergencialMapper.java
@@ -1,8 +1,8 @@
 package br.org.rapeifes.mapper;
 
+import br.org.rapeifes.domain.Aluno;
+import br.org.rapeifes.domain.ContatoEmergencial;
 import br.org.rapeifes.dto.ContatoEmergencialDTO;
-import br.org.rapeifes.model.Aluno;
-import br.org.rapeifes.model.ContatoEmergencial;
 import org.mapstruct.*;
 
 @Mapper(componentModel = "spring")
@@ -16,7 +16,7 @@ public interface ContatoEmergencialMapper extends GenericMapper<ContatoEmergenci
     @Mapping(target = "aluno", expression = "java(fromAlunoId(dto.getAlunoId()))")
     ContatoEmergencial toEntity(ContatoEmergencialDTO dto);
 
-    default Aluno fromAlunoId(Long id) {
+    default Aluno fromAlunoId(Integer id) {
         if (id == null) return null;
         return Aluno.builder().id(id).build();
     }
