+++ backend/src/main/java/br/org/rapeifes/repository/AlunoRepository.java
@@ -1,7 +1,7 @@
 package br.org.rapeifes.repository;
 
-import br.org.rapeifes.model.Aluno;
+import br.org.rapeifes.domain.Aluno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlunoRepository extends JpaRepository<Aluno, Long> {
+public interface AlunoRepository extends JpaRepository<Aluno, Integer> {
 }
