+++ backend/src/main/java/br/org/rapeifes/repository/ContatoEmergencialRepository.java
@@ -1,7 +1,7 @@
 package br.org.rapeifes.repository;
 
-import br.org.rapeifes.model.ContatoEmergencial;
+import br.org.rapeifes.domain.ContatoEmergencial;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContatoEmergencialRepository extends JpaRepository<ContatoEmergencial, Integer> {
+public interface ContatoEmergencialRepository extends JpaRepository<ContatoEmergencial, Long> {
 }
